worker/requirements.txt
psycopg2-binary==2.9.3
requests==2.26.0
GitPython==3.1.24
